pyproject.toml
[project]
name = "aisha"
version = "0.1.0"
description = "A small, opinionated AI assistant with memory, tools, and pluggable channels (Slack, WhatsApp, Telegram)."
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
dependencies = [
    "python-dotenv>=1.0.0",
    "chromadb>=0.4.22",
    "sentence-transformers>=2.2.2",
    "python-pptx>=0.6.23",
    "python-docx>=1.1.0",
    "reportlab>=4.0",
]

[project.optional-dependencies]
slack = ["websocket-client>=1.7.0"]
dev = ["pytest>=8.0.0", "ruff>=0.5.0"]

[project.scripts]
aisha = "aisha.__main__:main"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["aisha*"]

[tool.ruff]
line-length = 100
target-version = "py310"
